✨ feat: add blog functionality This commit adds functionality to the blog section of the website. The changes include: - Fixing a typo in the Header component by changing the path of the "Blog" link from "/blog" to "/blogs" - Adding a new component, PostCard, to display individual blog posts on the blogs page - Updating the next.config.js file to include support for swcMinify and the experimental appDir option, and to disable the fallback.fs option for non-server webpack builds - Adding two new API routes, /api/blogs and /api/blogs/[slug], to retrieve blog post data from markdown files - Adding two new pages, /blogs and /blogs/[slug], to display the list of blog posts and individual blog posts, respectively - Adding new dependencies to the package.json file, including @tailwindcss/typography, gray-matter, and markdown-to-jsx
📝 docs(posts): add blog post about using GPT-3 to generate poetry
A new blog post has been added about using GPT-3 to generate poetry and other creative content. The post explains what GPT-3 is and how it can be used for creative writing. It also includes a detailed description of the experiments conducted to test GPT-3's ability to generate poetry and stories. The post includes code snippets and images to illustrate the results of the experiments. Additionally, the post includes a link to the full set of results on GitHub. Finally, the post concludes with a wrap-up of the experiments and a link to a video covering the same content.

🖼️ feat(posts): add images to blog post about using GPT-3 to generate poetry
Two new images have been added to the blog post about using GPT-3 to generate poetry. The images are located in the public/images/posts directory and are named img-chatg